DataViews

Documentation for DataViews.

DataViews.AbstractIterator

DataViews.DataLoader

DataViews.JoinedView

DataViews.MappedView

DataViews.ObsView

DataViews.ZippedView

DataViews.dropobs

DataViews.filterobs

DataViews.getobs

DataViews.kfolds

DataViews.mapobs

DataViews.numobs

DataViews.obsview

DataViews.ones_like

DataViews.putobs

DataViews.repeatobs

DataViews.rmobs

DataViews.sampleobs

DataViews.shuffleobs

DataViews.splitobs

DataViews.stackobs

DataViews.takeobs

DataViews.unzip

DataViews.zeros_like

DataViews.zipobs

DataViews.AbstractIterator — Type

Super type of all iterators.

DataViews.DataLoader — Type

DataLoader(data; batchsize=1, partial=true, shuffle=false, parallel=true, rng=Random.default_rng())

An object that iterates over mini-batches of data, each mini-batch containing batchsize observations (except possibly the last one).

Takes as input a single data array, or in general any data object that implements the numobs and getobs methods.

The last dimension in each array is the observation dimension.

DataViews.JoinedView — Type

JoinedView(data...)

An object that iterates over each element in the iterators given by data as if they were concatenated into a single list.

DataViews.MappedView — Type

MappedView(f, data)

An iterator which lazily applies f to each element in data when requested.

DataViews.ObsView — Type

ObsView(data, indices)

Construct an iterator over the elements specified by indices in data.

DataViews.ZippedView — Type

ZippedView(data...)

Construct an iterator that zips each element of the given subiterators into a Tuple.

DataViews.dropobs — Method

dropobs(data, obs::AbstractVector{Int})

Remove all observations from data whose index corresponds to those in obs.

DataViews.filterobs — Method

filterobs(f, data)

Remove all observations from data for which f returns false.

DataViews.getobs — Method

getobs(data, [idx])

Return the observations corresponding to the observation index idx. Note that idx can be any type as long as data has defined getobs for that type. If idx is not provided, then materialize all observations in data.

Returns data[idx] if data does not have getobs defined.

Authors of custom data containers should implement Base.getindex for their type instead of getobs. getobs should only be implemented for types where there is a difference between getobs and Base.getindex (such as multi-dimensional arrays).

The returned observation(s) should be in the form intended to be passed as-is to some learning algorithm. There is no strict interface requirement on how this "actual data" must look like. Every author behind some custom data container can make this decision themselves. The output should be consistent when idx is a scalar vs vector.

DataViews.kfolds — Function

kfolds(data, k = 5)
kfolds(data::Integer, k = 5)
kfolds(data::AbstractVector{<:Integer}, k = 5)

Compute the train/validation splits for k repartitions of n observations, and return them as a vector of (train, validation) pairs. If data is an Integer, then a vector containing the indices 1:n will be materialized and partitioned. If data is an iterable, a lazy ObsView will be constructed for each fold.

DataViews.mapobs — Method

mapobs(f, data)

Lazily apply the function f to each element in data.

DataViews.numobs — Method

numobs(data)

Return the total number of observations contained in data.

Returns length(data) if data does not have numobs defined.

Arrays are a special case where numobs returns the size of the last dimension.

Authors of custom data containers should implement Base.length for their type instead of numobs. numobs should only be implemented for types where there is a difference between numobs and Base.length (such as multi-dimensional arrays).

DataViews.obsview — Method

obsview(data, indices::AbstractVector{<:Integer})

Construct a lazy view of data at the specified indices.

DataViews.ones_like — Method

ones_like(x::AbstractArray)

Construct an array of ones with the same size and element type as x.

DataViews.putobs — Method

putobs(x::AbstractArray)

Add an N+1 obervation dimension of size 1 to the tensor x.

DataViews.repeatobs — Method

repeatobs(data, n::Int)

Create a new view which iterates over every element in data n times.

DataViews.rmobs — Method

rmobs(x::AbstractArray)

Remove the observation dimension from the tensor x.

DataViews.sampleobs — Method

sampleobs([rng=default_rng()], data, n)

Randomly sample n elements from data without replacement. rng may be optionally provided for reproducible results.

DataViews.shuffleobs — Method

shuffleobs([rng=default_rng()], data)

Randomly shuffle the elements of data. Provide rng for reproducible results.

DataViews.splitobs — Method

splitobs([rng=default_rng()], data::Int; kw...)
splitobs([rng=default_rng()], data::AbstractVector{Int}; kw...)
splitobs([rng=default_rng()], data; at=0.8, shuffle=true)

Return a set of indices that splits the given observations according to the given break points.

Arguments

data: Any type that implements either Base.length() or numobs. Alternatively, can be

either an AbstractVector of indices or an Int indicating the number of observations.

at: The fractions at which to split data.

shuffle: If true, shuffles the indices before splitting.

Example

julia> splitobs(1:100, at=(0.7, 0.2), shuffle=false)
3-element Vector{Vector{Int64}}:
 [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  61, 62, 63, 64, 65, 66, 67, 68, 69, 70]
 [71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90]
 [91, 92, 93, 94, 95, 96, 97, 98, 99, 100]

DataViews.stackobs — Method

stackobs(x...)

Stack the elements in x as if they were observations in a batch. If x is an AbstractArray, elements will be concatenated along the Nth dimension. Other data types will simply be placed in a Vector in the same order as they were received. Special attention is paid to a collection of Tuples, where each tuple represents a single observation, such as a feature/label pair. In this case, the tuples will be unzipped and have their constituent elements stacked as usual.

Example

julia> stackobs(1, 2, 3, 4, 5)
5-element Vector{Int64}:
 1
 2
 3
 4
 5

julia> stackobs((1, :a), (2, :b), (3, :c))
([1, 2, 3], [:a, :b, :c])

julia> stackobs([rand(256, 256, 3, 1) for _ in 1:10]...) |> size
(256, 256, 3, 10)

julia> xs = [rand(256, 256, 3, 1) for _ in 1:10];

julia> ys = [rand(256, 256, 1, 1) for _ in 1:10];

julia> data = collect(zip(xs, ys));

julia> stackobs(data...) .|> size
((256, 256, 3, 10), (256, 256, 1, 10))

DataViews.takeobs — Method

takeobs(data, obs::AbstractVector{Int})

Take all observations from data whose index corresponds to obs while removing everything else.

DataViews.unzip — Method

unzip(x::AbstractVector{<:Tuple})

The inverse of zip.

Example

julia> zip([1, 2, 3], [:a, :b, :c]) |> collect |> unzip
([1, 2, 3], [:a, :b, :c])

DataViews.zeros_like — Method

zeros_like(x::AbstractArray)

Construct an array of zeros with the same size and element type as x.

DataViews.zipobs — Method

zipobs(data...)

Create a new iterator where the elements of each iterator in data are returned as a tuple.

Example

julia> zipobs(1:5, 41:45, [:a, :b, :c, :d, :e]) |> collect
5-element Vector{Any}:
 (1, 41, :a)
 (2, 42, :b)
 (3, 43, :c)
 (4, 44, :d)
 (5, 45, :e)